Add personal access token authentication
Introduces the bare minimum of an API authentication workflow. Needs
improvements such as hiding the token after verifying it (so it's not
just sitting on-screen in plain text), and possibly storing it in the OS
keychain using something like `keytar` so the user doesn't have to
generate new tokens each time the app is run.

The workflow has some basic error handling, and clears or resets page
elements back to their default values so the user doesn't get stuck in
an error loop.